+++ CHANGELOG.md
@@ -13,6 +13,7 @@ FEATURES:
 
 IMPROVEMENTS:
 
+- Add account pre-check for common bandwidth package to avoid known error [GH-877]
 - Make dns resource code more standard [GH-876]
 - Improve dns resources' testcases [GH-859]
 - Add client token for vpn services [GH-855]